player can now control the tank's turret

+++ jni/Match.cpp
@@ -108,7 +108,7 @@ void Match::recv_data(State &state){
 			player.yv=(int)ntohl(tch.state[(i*SERVER_STATE_FIELDS)+SERVER_STATE_YV])/FLOAT_MULTIPLIER;
 			player.x=(int)ntohl(tch.state[(i*SERVER_STATE_FIELDS)+SERVER_STATE_XPOS])/FLOAT_MULTIPLIER;
 			player.y=(int)ntohl(tch.state[(i*SERVER_STATE_FIELDS)+SERVER_STATE_YPOS])/FLOAT_MULTIPLIER;
-			player.turret.rot=ntohl(tch.state[(i*SERVER_STATE_FIELDS)+SERVER_STATE_ANGLE])/FLOAT_MULTIPLIER;
+			player.turret.rot=(int)ntohl(tch.state[(i*SERVER_STATE_FIELDS)+SERVER_STATE_ANGLE])/FLOAT_MULTIPLIER;
 			player.colorid=ntohl(tch.state[(i*SERVER_STATE_FIELDS)+SERVER_STATE_COLORID]);
 			player.cue_fire=ntohl(tch.state[(i*SERVER_STATE_FIELDS)+SERVER_STATE_FIRE]);
